refactor(occlusion): move SoundPhysicsCompat to Util, show state in /rpvc info
- Relocate SoundPhysicsCompat from Audio/Effects to Util (it's a data
  provider, not an OpenAL EFX wrapper) to match their layering
- /rpvc info now reports the use-sound-physics-adapted state
